chore: moved intro to onboarding folder
